ParetoTest {quantreg}	R Documentation

Estimation and Inference on the Pareto Tail Exponent for Linear Models

Description

Estimation and inference about the tail behavior of the response in linear models are based on the adaptation of the univariate Hill (1975) and Pickands (1975) estimators for quantile regression by Chernozhukov, Fernandez-Val and Kaji (2018).

Usage

ParetoTest(formula, tau = 0.1, data = NULL, flavor = "Hill", m = 2, cicov = .9, ...) 

Arguments

formula
a formula specifying the model to fit by rq

tau
A threshold on which to base the estimation

data
a data frame within which to interpret the formula

flavor
Currently limited to either "Hill" or "Pickands"

m
a tuning parameter for the Pickands method .

cicov
Desired coverage probability of confidence interval.

...
other arguments to be passed to summary.rq. by default the summary method is the usual xy bootstrap, with B = 200 replications.

Value

an object of class ParetoTest is returned containing:

z
A named vector with components: the estimate, a bias corrected estimate, a lower bound of the confidence interval, an upper bound of the confidence interval, and a Bootstrap Standard Error estimate.

tau
The tau threshold used to compute the estimate

References

Chernozhukov, Victor, Ivan Fernandez-Val, and Tetsuya Kaji, (2018) Extremal Quantile Regression, in Handbook of Quantile Regression, Eds. Roger Koenker, Victor Chernozhukov, Xuming He, Limin Peng, CRC Press.

Hill, B. M. (1975). A simple general approach to inference about the tail of a distribution. The Annals of Statistics 3(5), 1163-1174.

Pickands, J. (1975). Statistical inference using extreme order statistics. The Annals of Statistics 3(1), 119-131.

Examples

n = 500
x = rnorm(n)
y = x + rt(n,2)
Z = ParetoTest(y ~ x, .9, flavor = "Pickands")

[Package quantreg version 5.97 Index]