akj {quantreg}	R Documentation

Density estimation using adaptive kernel method

Description

univariate adaptive kernel density estimation a la Silverman. As used by Portnoy and Koenker (1989)

Usage

akj(x, z, p, h, alpha, kappa, iker1, iker2)

Arguments

x 	 points used for centers of kernel assumed to be sorted

z 	 points at which density is calculated; default to seq( min(x), max(x), 2*length(x) )

p 	 vector of probabilities associated with x's; default to 1/len(x) for each x.

h 	 initial window size (overall); default to Silverman's normal reference

alpha 	 a sensitivity parameter that determines the sensitivity of the local bandwidth to variations in the pilot density; default to .5

kappa 	 constant determining initial (default) window width

iker1 	 kernel indicator, 0 for normal kernel (default) while 1 for cauchy kernel

iker2 	 xxx

Value

a R structure is returned

dens 	 the vector of estimated density

psi 	 a vector of psi=-f'/f function

score 	 a vector of score (f'/f)^2-f''/f function

h 	 same as the input argument h

References

Portnoy, S and R Koenker, (1989) Adaptive L Estimation of Linear Models, Annals, 17, 362-81. Silverman, B. (1986) Density Estimation, pp100-104.

[Package quantreg version 3.82 Index]