boot.rq {quantreg}	R Documentation

Bootstrapping Quantile Regression

Description

These functions can be used to construct standard errors, confidence intervals and tests of hypotheses regarding quantile regression models.

Usage

boot.rq(x, y, tau = 0.5, R = 200, bsmethod = "xy", mofn = length(y), ...)

Arguments

x 	 The regression design matrix

y 	 The regression response vector

tau 	 The quantile of interest

R 	 The number of bootstrap replications

bsmethod 	 The method to be employed. There are (as yet) three options: method = "xy" uses the xy-pair method, and method = "pwy" uses the method of Parzen, Wei and Ying (1994) method = "mcmb" uses the Markov chain marginal bootstrap of He and Hu (2002) and Kocherginsky, He and Mu (2003).

mofn 	 optional argument for the bootstrap method "xy" that permits subsampling (m out of n) bootstrap. Obviously mofn should be substantially larger than the column dimension of x, and should be less than the sample size.

... 	 Optional arguments to control bootstrapping

Details

Their are several refinements that are still unimplemented. Percentile methods should be incorporated, and extensions of the methods to be used in anova.rq should be made.

Value

A matrix of dimension R by p is returned with the R resampled estimates of the vector of quantile regression parameters. When mofn < n for the "xy" method this matrix has been deflated by the fact sqrt(m/n)

Author(s)

Roger Koenker (and Xuming He and M. Kocherginsky for the mcmb code)

References

[1] Koenker, R. W. (1994). Confidence Intervals for regression quantiles, in P. Mandl and M. Huskova (eds.), Asymptotic Statistics, 349–359, Springer-Verlag, New York.

[2] Kocherginsky, M., He, X. and Mu, Y. (2003). Practical Confidence Intervals for Regression Quantiles. Preprint.

[3] He, X. and Hu, F. (2002). Markov Chain Marginal Bootstrap. Journal of the American Statistical Association , Vol. 97, no. 459, 783-795.

[4] Parzen, M. I., L. Wei, and Z. Ying (1994): A resampling method based on pivotal estimating functions,'' Biometrika, 81, 341–350.

See Also

summary.rq

Examples

y <- rnorm(50)
x <- matrix(rnorm(100),50)
fit <- rq(y~x,tau = .4)
summary(fit,se = "boot", bsmethod= "xy")
summary(fit,se = "boot", bsmethod= "pwy")
#summary(fit,se = "boot", bsmethod= "mcmb")

[Package quantreg version 3.82 Index]