engel {quantreg}	R Documentation

Engel Data

Description

Engel Food Expenditure Data used in Koenker and Bassett(1982). This is a regression data set consisting of 235 observations on income and expenditure on food for Belgian working class households.

Usage

data(engel)

Format

A data frame containing 235 observations on 2 variables

x

"Annual Household Income in Belgian Francs"

y

"Annual Household Food Expenditure in Belgian Francs"

References

Koenker, R. and Bassett, G (1982) Robust Tests of Heteroscedasticity based on Regression Quantiles, Econometrica, 50,43–61.

[Package quantreg version 3.82 Index]