nlrq.control {quantreg}	R Documentation

Set control parameters for nlrq

Description

Set algorithmic parameters for nlrq (nonlinear quantile regression function)

Usage

nlrq.control(maxiter=100, k=2, InitialStepSize = 1, big=1e+20, eps=1e-07, beta=0.97)

Arguments

maxiter
maximum number of allowed iterations

k
the number of iterations of the Meketon algorithm to be calculated in each step, usually 2 is reasonable, occasionally it may be helpful to set k=1

InitialStepSize
Starting value in optim to determine the step length of iterations. The default value of 1 is sometimes too optimistic. In such cases, the value 0 forces optim to just barely stick its toe in the water.

big
a large scalar

eps
tolerance for convergence of the algorithm

beta
a shrinkage parameter which controls the recentering process in the interior point algorithm.

See Also

nlrq

[Package quantreg version 5.97 Index]