predict.qss {quantreg}	R Documentation

Predict based on nonparametric quantile regression smoothing spline component

Description

Additive models for nonparametric quantile regression using total variation penalty methods can be fit with the rqss function. Univarariate and bivariate components can be predicted using these functions.

Usage

predict.qss1(object, newdata, ...)
predict.qss2(object, newdata, ...)

Arguments

object 	 is a fitted object produced by rqss

newdata 	 a data frame describing the observations at which prediction is to be made

... 	 optional arguments

Details

For both univariate and bivariate prediction linear interpolation is done. In the bivariate case, this involves computing barycentric coordinates of the new points relative to their enclosing triangles.

Value

A list consisting of x and y components in the case of qss1, and a list consisting of x, y, and z components in the case of qss2. In the former case the y component constitutes the predictions at x, and in the latter, z is the vector of the predictions at the points (x,y).

Author(s)

R. Koenker

See Also

rqss

Examples

[Package quantreg version 3.82 Index]