ranks {quantreg}	R Documentation

Quantile Regression Ranks

Description

Function to compute ranks from the dual (regression rankscore) process.

Usage

ranks(v, score="wilcoxon", tau=0.5)

Arguments

v 	 object of class "rq.process" generated by rq()

score 	 The score function desired. Currently implemented score functions are "wilcoxon", "normal", and "sign" which are asymptotically optimal for the logistic, Gaussian and Laplace location shift models respectively. The "normal" score function is also sometimes called van der Waerden scores. Also implemented are the "tau" which generalizes sign scores to an arbitrary quantile, and "interquartile" which is appropriate for tests of scale shift.

tau 	 the optional value of tau if the "tau" score function is used.

Details

See GJKP(1993) for further details.

Value

The function returns two components. One is the ranks, the other is a scale factor which is the L_2 norm of the score function. All score functions should be normalized to have mean zero.

References

Gutenbrunner, C., J. Jureckova, Koenker, R. and Portnoy, S. (1993) Tests of linear hypotheses based on regression rank scores, Journal of Nonparametric Statistics, (2), 307–331.

See Also

rq, rq.test.rank anova.rq

Examples

data(stackloss)
ranks(rq(stack.loss ~ stack.x, tau=-1))

[Package quantreg version 3.82 Index]