residuals.nlrq {quantreg}	R Documentation

Return residuals of an nlrq object

Description

Set algorithmic parameters for nlrq (nonlinear quantile regression function)

Usage

residuals.nlrq(object, type = c("response", "rho"), ...) 

Arguments

object 	 an `nlrq' object as returned by function `nlrq'

type 	 the type of residuals to return: "response" is the distance between observed and predicted values; "rho" is the weighted distance used to calculate the objective function in the minimisation algorithm as tau * pmax(resid, 0) + (1 - tau) * pmin(resid, 0), where resid are the simple residuals as above (with type="response").

... 	 further arguments passed to or from other methods.

See Also

nlrq

[Package quantreg version 3.82 Index]