rq.fit.pfn {quantreg}	R Documentation

Preprocessing Algorithm for Quantile Regression

Description

A preprocessing algorithm for the Frisch Newton algorithm for quantile regression. This is one possible method for rq().

Usage

rq.fit.pfn(x, y, tau=0.5, Mm.factor=0.8, max.bad.fixups=3, eps=1e-06)

Arguments

x
design matrix usually supplied via rq()

y
response vector usually supplied via rq()

tau
quantile of interest

Mm.factor
constant to determine sub sample size m

max.bad.fixups
number of allowed mispredicted signs of residuals

eps
convergence tolerance

Details

Preprocessing algorithm to reduce the effective sample size for QR problems with (plausibly) iid samples. The preprocessing relies on subsampling of the original data, so situations in which the observations are not plausibly iid, are likely to cause problems. The tolerance eps may be relaxed somewhat.

Value

Returns an object of type rq

Author(s)

Roger Koenker <rkoenker@uiuc.edu>

References

Portnoy and Koenker, Statistical Science, (1997) 279-300

See Also

rq

[Package quantreg version 5.97 Index]