rq.fit.ppro {quantreg}	R Documentation

Preprocessing fitting method for QR

Description

Preprocessing method for fitting quantile regression models that exploits the fact that adjacent tau's should have nearly the same sign vectors for residuals.

Usage

rq.fit.ppro(x, y, tau, weights = NULL, Mm.factor = 0.8, eps = 1e-06, ...)

Arguments

x
Design matrix

y
Response vector

tau
quantile vector of interest

weights
case weights

Mm.factor
constant determining initial sample size

eps
Convergence tolerance

...
Other arguments

Details

See references for further details.

Value

Returns a list with components:

coefficients
Matrix of coefficient estimates

residuals
Matrix of residual estimates

rho
vector of objective function values

weights
vector of case weights

Author(s)

Blaise Melly and Roger Koenker

References

Chernozhukov, V. I. Fernandez-Val and B. Melly, Fast Algorithms for the Quantile Regression Process, 2020, Empirical Economics.,

Portnoy, S. and R. Koenker, The Gaussian Hare and the Laplacian Tortoise, Statistical Science, (1997) 279-300

See Also

rq.fit.pfn, boot.rq.pxy

[Package quantreg version 5.97 Index]