rq.object {quantreg}	R Documentation

Linear Quantile Regression Object

Description

These are objects of class "rq". They represent the fit of a linear conditional quantile function model.

Details

The coefficients, residuals, and effects may be extracted by the generic functions of the same name, rather than by the $ operator. For pure rq objects this is less critical than for some of the inheritor classes. Note that the extractor function coef returns a vector with missing values omitted.

Generation

This class of objects is returned from the rq function to represent a fitted linear quantile regression model.

Methods

The "rq" class of objects has methods for the following generic functions: coef, effects , formula , labels , model.frame , model.matrix , plot , predict , print , print.summary , residuals , summary

Structure

The following components must be included in a legitimate rq object.

coefficients

the coefficients of the quantile regression fit. The names of the coefficients are the names of the single-degree-of-freedom effects (the columns of the model matrix). If the model was fitted by method "br" with ci=TRUE, then the coefficient component consists of a matrix whose first column consists of the vector of estimated coefficients and the second and third columns are the lower and upper limits of a confidence interval for the respective coefficients.

residuals

the residuals from the fit.

contrasts

a list containing sufficient information to construct the contrasts used to fit any factors occurring in the model. The list contains entries that are either matrices or character vectors. When a factor is coded by contrasts, the corresponding contrast matrix is stored in this list. Factors that appear only as dummy variables and variables in the model that are matrices correspond to character vectors in the list. The character vector has the level names for a factor or the column labels for a matrix.

model

optionally the model frame, if model=TRUE.

x

optionally the model matrix, if x=TRUE.

y

optionally the response, if y=TRUE.

See Also

rq, coefficients.

[Package quantreg version 3.82 Index]