rqProcess {quantreg}	R Documentation

Compute Quantile Regression Process

Description

Computes the quantile regression process for the model specified by the formula on the grid specified by the taus argument. Intended for use in khmaladze.test.

Usage

rqProcess(formula, data, taus = seq(0.2, 0.8, by = 0.1))

Arguments

formula 	 model formula

data 	 data frame to be used to interpret formula

taus 	 quantiles at which the process is to be evaluated

Details

The process computes point estimates and also returns an array with the J and Hinv matrices needed to compute standardizations.

Value

coefs 	 Point estimates of the coefs of the model

J 	 J matrix of the usual qr sandwich formula, without any tau(1-tau) factor

Hinv 	 Hinv matrices of the usual qr sandwich formula, this is an array of dimension (p,p,m)

Author(s)

R. Koenker

See Also

table.rq, khmaladze.test

Examples

[Package quantreg version 3.82 Index]