rrs.test {quantreg}	R Documentation

Quantile Regression Rankscore Test

Description

Function to compute regression rankscore test of a linear hypothesis based on the dual quantile regression process. A test of the hypothesis, is carried out by estimating the restricted model and constructing a test based on the dual process under the restricted model. The details of the test are described in GJKP(1993). The test has a Rao-score, Lagrange-multiplier interpretation since in effect it is based on the value of the gradient of unrestricted quantile regression problem evaluated under the null. This function will eventually be superseded by a more general anova() method for rq.

Usage

rrs.test(x0, x1, y, v, score="wilcoxon")

Arguments

x0 	 the matrix of maintained regressors, a column of ones is appended automatically.

x1 	 matrix of covariates under test.

y 	 response variable, may be omitted if v is provided.

v 	 object of class "rq.process" generated e.g. by rq(y ~ x0, tau=-1)

score 	 Score function for test (see ranks)

Details

See GJKP(1993)

Value

Test statistic sn is asymptotically Chi-squared with rank(X1) dfs. The vector of ranks is also returned as component rank.

References

[1] Gutenbrunner, C., Jureckova, J., Koenker, R. and Portnoy, S. (1993) Tests of linear hypotheses based on regression rank scores. Journal of Nonparametric Statistics, (2), 307-331.

[2] Koenker, R. W. and d'Orey (1994). Remark on Alg. AS 229: Computing dual regression quantiles and regression rank scores. Applied Statistics, 43, 410-414.

See Also

rq, ranks

Examples

# Test that covariates 2 and 3 belong in stackloss model using Wilcoxon scores.
data(stackloss)
rrs.test(stack.x[,1], stack.x[,2:3], stack.loss)

[Package quantreg version 3.82 Index]