sfn.control {quantreg}	R Documentation

Set Control Parameters for Sparse Fitting

Description

Auxiliary function for setting storage dimensions and other parameters rq.fit.sfn[c]

Usage

sfn.control(nsubmax = NULL, tmpmax = NULL, nnzlmax = NULL, cachsz = 64, 
	small = 1e-06, maxiter = 100, warn.mesg = TRUE)

Arguments

nsubmax
upper bound for dimension of lindx

tmpmax
upper bound for dimension of tmpvec

nnzlmax
upper bound for non-zero entries of L stored in lnz, including diagonal

cachsz
size of cache in kbytes on target machine

small
convergence tolerance for interior point algorithm

maxiter
maximal number of interior point iterations.

warn.mesg
logical flag controlling printing of warnings.

Details

Sparse fitting requires a number of temporary storage arrays whose size depends on problem specific features in somewhat mysterious ways, parameters controlling these sizes and some other fitting aspects can be controlled by specifying elements of this control object.

Value

List with components named as the arguments given above.

Author(s)

Roger Koenker

See Also

See Also rq.fit.sfn

[Package quantreg version 5.97 Index]