srisk {quantreg}	R Documentation

Markowitz (Mean-Variance) Portfolio Optimization

Description

This function estimates optimal mean-variance portfolio weights from a matrix of historical or simulated asset returns.

Usage

srisk(x, mu = 0.07, lambda = 1e+08, alpha = 0.1, eps = 1e-04)

Arguments

x
Matrix of asset returns

mu
Required mean rate of return for the portfolio

lambda
Lagrange multiplier associated with mean return constraint

alpha
Choquet risk parameter, unimplemented

eps
tolerance parameter for mean return constraint

Details

The portfolio weights are estimated by solving a constrained least squares problem.

Value

pihat
Optimal portfolio weights

muhat
Mean return in sample

sighat
Standard deviation of returns in sample

Author(s)

R. Koenker

See Also

qrisk

[Package quantreg version 5.97 Index]