summary.crq {quantreg}	R Documentation

Summary methods for Censored Quantile Regression

Description

Returns a summary object for a censored quantile regression fit. A null value will be returned if printing is invoked.

Usage

## S3 method for class 'crq'
summary(object, taus = 1:4/5, alpha = .05, se="boot", covariance=TRUE,  ...)
## S3 method for class 'summary.crq'
print(x, digits = max(5, .Options$digits - 2), ...)
## S3 method for class 'summary.crqs'
print(x,  ...)
## S3 method for class 'summary.crqs'
plot(x, nrow = 3, ncol = 3, CoxPHit = NULL,  ...)

Arguments

object
An object of class "crq" produced by a call to crq().

taus
Quantiles to be summarized. This should be a vector of length greater than one.

x
An object of class "crq" produced by a call to crq().

se
specifies the method used to compute standard standard errors. but the only available method (so far) is "boot". Further arguments to boot.crq and boot.rq can be passed with the ... argument.

covariance
logical flag to indicate whether the full covariance matrix of the estimated parameters should be returned.

nrow
Number of rows of the plot layout.

ncol
Number of columns of the plot layout.

alpha
Confidence level for summary intervals.

digits
Number of digits to be printed in summary display.

CoxPHit
An object of class coxph produced by coxph.

...
Optional arguments to summary, e.g. to specify bootstrap methods sample sizes, etc. see boot.rq and boot.crq

Details

For the Powell method the resampling strategy used by the se = "boot" method is based on the Bilias, Chen and Ying (2000) proposal. For the Portnoy and Peng-Huang methods the bootstrapping is by default actually based on a delete-d jackknife, as described in Portnoy (2013), but resampling xy pairs using either conventional multinomial resampling or using exponential weighting as in Bose and Chatterjee (2003) can be used by specifying the bmethod argument. Note that the default number of replications is set at R = 100 a value that is obviously too small for most applications. This is done merely to speed up the examples in the documentation and facilitate testing. Larger, more appropriate values of R can be passed to the bootstrapping functions via the ... argument of the summary method. It is important to recognize that when some of the bootstrap replications are NA they are simply ignored in the computation of the confidence bands and standard errors as currently reported. The number of these NAs is returned as part of the summary.crq object, and when printed is also reported.

Value

For method "Powell" an object of class summary.crq is returned with the following components:

coefficients
a p by 4 matrix consisting of the coefficients, their estimated standard errors, their t-statistics, and their associated p-values.

cov
the estimated covariance matrix for the coefficients in the model, provided that covariance = TRUE appears in the calling sequence.

rdf
the residual degrees of freedom

tau
the quantile estimated

For the other methods an object of class summary.crq is returned with the following components:

coefficients
a list of p by 6 matrix consisting of the coefficients, upper and lower bounds for a (1-alpha) level confidence interval, their estimated standard errors, their t-statistics, and their associated p-values, one component for each element of the specified taus vector.

cov
the estimated covariance matrix for the coefficients in the model, provided that covariance = TRUE in the called sequence.

References

Bose, A. and S. Chatterjee, (2003) Generalized bootstrap for estimators of minimizers of convex functions, J. Stat. Planning and Inf, 117, 225-239.

Bilias, Y. Chen, S. and Z. Ying, (2000) Simple resampling methods for censored quantile regression, J. of Econometrics, 99, 373-386.

Portnoy, S. (2013) The Jackknife's Edge: Inference for Censored Quantile Regression, CSDA, forthcoming.

See Also

crq, QTECox

[Package quantreg version 5.94 Index]