summary.rq {quantreg}	R Documentation

Summary methods for Quantile Regression

Description

Returns a summary list for a quantile regression fit. A null value will be returned if printing is invoked.

Usage

summary.rq(object, se="nid", covariance=TRUE, hs = TRUE,  ...)
summary.rqs(object, ...)

Arguments

object 	 This is an object of class "rq" or "rqs" produced by a call to rq(), depending on whether one or more taus are specified.

se 	 specifies the method used to compute standard standard errors. There are currently five available methods:

"rank" which produces confidence intervals for the estimated parameters by inverting a rank test as described in Koenker (1994). The default option assumes that the errors are iid, while the option iid = FALSE implements the proposal of Koenker Machado (1999).

"iid" which presumes that the errors are iid and computes an estimate of the asymptotic covariance matrix as in KB(1978).

"nid" which presumes local (in tau) linearity (in x) of the the conditional quantile functions and computes a Huber sandwich estimate using a local estimate of the sparsity.

"ker" which uses a kernel estimate of the sandwich as proposed by Powell(1990).

"boot" which implements one of several possible bootstrapping alternatives for estimating standard errors.

covariance 	 logical flag to indicate whether the full covariance matrix of the estimated parameters should be returned.

hs 	 Use Hall Sheather bandwidth for sparsity estimation If false revert to Bofinger bandwidth.

... 	 Optional arguments to summary, e.g. bsmethod to use bootstrapping. see boot.rq

Details

When the default summary method is used, it tries to estimate a sandwich form of the asymptotic covariance matrix and this involves estimating the conditional density at each of the sample observations, negative estimates can occur if there is crossing of the neighboring quantile surfaces used to compute the difference quotient estimate. If the number of these is large relative to the sample size it is sometimes an indication that some additional nonlinearity in the covariates would be helpful, for instance quadratic effects.

Value

a list is returned with the following components, when object is of class "rqs" then there is a list of such lists.

coefficients 	 a p by 4 matrix consisting of the coefficients, their estimated standard errors, their t-statistics, and their associated p-values.

cov 	 the estimated covariance matrix for the coefficients in the model, provided that cov=TRUE in the called sequence.

Hinv 	 inverse of the estimated Hessian matrix returned if cov=TRUE and se != "iid".

J 	 Outer product of gradient matrix returned if cov=TRUE and se != "iid". The Huber sandwich is cov = Hinv %*% J %*% Hinv.

References

Koenker, R. (2004) Quantile Regression.

See Also

rq bandwidth.rq

Examples

data(stackloss)
y <- stack.loss
x <- stack.x
summary(rq(y ~ x, method="fn")) # Compute se's for fit using "nid" method.
summary(rq(y ~ x, ci=FALSE),se="ker")
# default "br" alg, and compute kernel method se's

[Package quantreg version 3.82 Index]