boot.rq.pwxy {quantreg}	R Documentation

Preprocessing weighted bootstrap method

Description

Bootstrap method exploiting preprocessing strategy to reduce computation time for large problem. In contrast to boot.rq.pxy which uses the classical multinomial sampling scheme and is coded in R, this uses the exponentially weighted bootstrap scheme and is coded in fortran and consequently is considerably faster in larger problems.

Usage

boot.rq.pwxy(x, y, tau, coef, R = 200, m0 = NULL, eps = 1e-06, ...)

Arguments

x
Design matrix

y
response vector

tau
quantile of interest

coef
point estimate of fitted object

R
the number of bootstrap replications desired.

m0
constant to determine initial sample size, defaults to sqrt(n*p) but could use some further tuning...

eps
tolerance for convergence of fitting algorithm

...
other parameters not yet envisaged.

Details

The fortran implementation is quite similar to the R code for boot.rq.pxy except that there is no multinomial sampling. Instead rexp(n) weights are used.

Value

returns a list with elements:

coefficientsa matrix of dimension ncol(x) by R

nit a 5 by m matrix of iteration counts

info an m-vector of convergence flags

Author(s)

Blaise Melly and Roger Koenker

References

Chernozhukov, V. I. Fernandez-Val and B. Melly, Fast Algorithms for the Quantile Regression Process, 2019, arXiv, 1909.05782,

Portnoy, S. and R. Koenker, The Gaussian Hare and the Laplacian Tortoise, Statistical Science, (1997) 279-300

See Also

boot.rq.pxy

[Package quantreg version 5.94 Index]