boot.rq.pxy {quantreg}	R Documentation

Preprocessing bootstrap method

Description

Bootstrap method exploiting preprocessing strategy to reduce computation time for large problem.

Usage

boot.rq.pxy(x, y, s, tau = 0.5, coef, method = "fn", Mm.factor = 3)

Arguments

x
Design matrix

y
response vector

s
matrix of multinomial draws for xy bootstrap

tau
quantile of interest

coef
point estimate of fitted object

method
fitting method for bootstrap

Mm.factor
constant to determine initial sample size

Details

See references for further details.

Value

Returns matrix of bootstrap estimates.

Author(s)

Blaise Melly and Roger Koenker

References

Chernozhukov, V. I. Fernandez-Val and B. Melly, Fast Algorithms for the Quantile Regression Process, 2019, arXiv, 1909.05782,

Portnoy, S. and R. Koenker, The Gaussian Hare and the Laplacian Tortoise, Statistical Science, (1997) 279-300

See Also

rq.fit.ppro

[Package quantreg version 5.93 Index]