critval {quantreg}	R Documentation

Hotelling Critical Values

Description

Critical values for uniform confidence bands for rqss fitting

Usage

critval(kappa, alpha = 0.05, rdf = 0)

Arguments

kappa
length of the tube

alpha
desired non-coverage of the band, intended coverage is 1 - alpha

rdf
"residual" degrees of freedom of the fitted object. If rdf=0 then the Gaussian version of the critical value is computed, otherwise the value is based on standard Student t theory.

Details

The Hotelling tube approach to inference has a long and illustrious history. See Johansen and Johnstone (1989) for an overview. The implementation here is based on Sun and Loader (1994) and Loader's locfit package, although a simpler root finding approach is substituted for the iterative method used there. At this stage, only univariate bands may be constructed.

Value

A scalar critical value that acts as a multiplier for the uniform confidence band construction.

References

Hotelling, H. (1939): “Tubes and Spheres in $n$-spaces, and a class of statistical problems,” Am J. Math, 61, 440–460.

Johansen, S., I.M. Johnstone (1990): “Hotelling's Theorem on the Volume of Tubes: Some Illustrations in Simultaneous Inference and Data Analysis,” The Annals of Statistics, 18, 652–684.

Sun, J. and C.V. Loader: (1994) “Simultaneous Confidence Bands for Linear Regression and smoothing,” The Annals of Statistics, 22, 1328–1345.

See Also

plot.rqss

[Package quantreg version 5.94 Index]