latex.summary.rqs {quantreg}	R Documentation

Make a latex table from a table of rq results

Description

Produces a file with latex commands for a table of rq results.

Usage

## S3 method for class 'summary.rqs'
latex(x, transpose = FALSE, caption = "caption goes here.", 
		digits = 3, file = as.character(substitute(x)), ...)

Arguments

x
x is an object of class summary.rqs

transpose
if TRUE transpose table so that rows are quantiles and columns are covariates.

caption
caption for the table

digits
decimal precision of table entries.

file
name of file

...
optional arguments for latex.table

Details

Calls latex.table.

Value

Returns invisibly after writing the file.

Author(s)

R. Koenker

See Also

summary.rqs, latex.table

[Package quantreg version 5.93 Index]