lm.fit.recursive {quantreg}	R Documentation

Recursive Least Squares

Description

This function fits a linear model by recursive least squares. It is a utility routine for the khmaladzize function of the quantile regression package.

Usage

lm.fit.recursive(X, y, int=TRUE)

Arguments

X 	 Design Matrix

y 	 Response Variable

int 	 if TRUE then append intercept to X

Value

return p by n matrix of fitted parameters, where p. The ith column gives the solution up to "time" i.

Author(s)

R. Koenker

References

A. Harvey, (1993) Time Series Models, MIT

See Also

khmaladzize

[Package quantreg version 3.82 Index]