rq.fit.conquer {quantreg}	R Documentation

Optional Fitting Method for Quantile Regression

Description

This fitting method provides a link to the gradient descent for convolution smoothed quantile regression problem implemented in the conquer package of He et al (2020).

Usage

rq.fit.conquer (x, y, tau=0.5, kernel = c("Gaussian", "uniform",
    "parabolic", "triangular"), h = 0, tol = 1e-04,
    iteMax = 5000, ci = FALSE, alpha = 0.05, B = 200)

Arguments

x
design matrix usually supplied via rq(), expected to have a intercept as the first column

y
response vector usually supplied via rq()

tau
quantile of interest

kernel
A character string specifying the choice of kernel function. Default is "Gaussian". Other choices are "uniform", "parabolic" or "triangular".

h
The bandwidth parameter for kernel smoothing of the QR objective function. Default is max((log(n) + p) / n)^0.4, 0.05. The default is used if the input value is less than 0.05.

tol
Tolerance level of the gradient descent algorithm. The gradient descent algorithm terminates when the maximal entry of the gradient is less than "tol". Default is 1e-05.

iteMax
Maximum number of iterations. Default is 5000.

ci
A logical flag. Default is FALSE. If "ci = TRUE", then three types of confidence intervals (percentile, pivotal and normal) will be constructed via multiplier bootstrap. This option is subsumed in normal use by the summary.rq functionality.

alpha
Nominal level for confidence intervals, may be passed via the call to summary

B
Number of bootstrap replications. May be passed via summary.

Details

See documentation in the conquer package.

Value

Returns an object of class "rq".

References

Xuming He and Xiaoou Pan and Kean Ming Tan and Wen-Xin Zhou, (2020) conquer: Convolution-Type Smoothed Quantile Regression, https://CRAN.R-project.org/package=conquer

See Also

rq

[Package quantreg version 5.94 Index]