rq.fit.fnc {quantreg}	R Documentation

Quantile Regression Fitting via Interior Point Methods

Description

This is a lower level routine called by rq() to compute quantile regression methods using the Frisch-Newton algorithm. It allows the call to specify linear inequality constraints to which the fitted coefficients will be subjected. The constraints are assumed to be formulated as Rb >= r.

Usage

rq.fit.fnc(x, y, R, r, tau=0.5, beta=0.9995, eps=1e-06)

Arguments

x
The design matrix

y
The response vector

R
The matrix describing the inequality constraints

r
The right hand side vector of inequality constraints

tau
The quantile of interest, must lie in (0,1)

beta
technical step length parameter – alter at your own risk!

eps
tolerance parameter for convergence. In cases of multiple optimal solutions there may be some discrepancy between solutions produced by method "fn" and method "br". This is due to the fact that "fn" tends to converge to a point near the centroid of the solution set, while "br" stops at a vertex of the set.

Details

The details of the algorithm are explained in Koenker and Ng (2002). The basic idea can be traced back to the log-barrier methods proposed by Frisch in the 1950's for constrained optimization. But the current implementation is based on proposals by Mehrotra and others in the recent (explosive) literature on interior point methods for solving linear programming problems. See "rq" helpfile for an example. It is an open research problem to provide an inference apparatus for inequality constrained quantile regression.

Value

returns an object of class "rq", which can be passed to summary.rq to obtain standard errors, etc.

References

Koenker, R. and S. Portnoy (1997). The Gaussian Hare and the Laplacian Tortoise: Computability of squared-error vs. absolute-error estimators, with discussion, Statistical Science, 12, 279-300.

Koenker, R. and P. Ng(2005). Inequality Constrained Quantile Regression, Sankya, 418-440.

See Also

rq, rq.fit.br, rq.fit.pfn

[Package quantreg version 5.93 Index]