rq.fit.hogg {quantreg}	R Documentation

weighted quantile regression fitting

Description

Function to estimate a regression mmodel by minimizing the weighted sum of several quantile regression functions. See Koenker(1984) for an asymptotic look at these estimators. This is a slightly generalized version of what Zou and Yuan (2008) call composite quantile regression in that it permits weighting of the components of the objective function and also allows further linear inequality constraints on the coefficients.

Usage

rq.fit.hogg(x, y, taus = c(0.1, 0.3, 0.5), weights = c(0.7, 0.2, 0.1), 
	    R = NULL, r = NULL, beta = 0.99995, eps = 1e-06)

Arguments

x
design matrix

y
response vector

taus
quantiles getting positive weight

weights
weights assigned to the quantiles

R
optional matrix describing linear inequality constraints

r
optional vector describing linear inequality constraints

beta
step length parameter of the Frisch Newton Algorithm

eps
tolerance parameter for the Frisch Newton Algorithm

Details

Mimimizes a weighted sum of quantile regression objective functions using the specified taus. The model permits distinct intercept parameters at each of the specified taus, but the slope parameters are constrained to be the same for all taus. This estimator was originally suggested to the author by Bob Hogg in one of his famous blue notes of 1979. The algorithm used to solve the resulting linear programming problems is either the Frisch Newton algorithm described in Portnoy and Koenker (1997), or the closely related algorithm described in Koenker and Ng(2002) that handles linear inequality constraints. See qrisk for illustration of its use in portfolio allocation.

Linear inequality constraints of the form Rb ≥q r can be imposed with the convention that b is a m+p where m is the length(taus) and p is the column dimension of x without the intercept.

Value

coefficients
estimated coefficients of the model

Author(s)

Roger Koenker

References

Zou, Hui and and Ming Yuan (2008) Composite quantile regression and the Oracle model selection theory, Annals of Statistics, 36, 1108–11120.

Koenker, R. (1984) A note on L-estimates for linear models, Stat. and Prob Letters, 2, 323-5.

Portnoy, S. and Koenker, R. (1997) The Gaussian Hare and the Laplacean Tortoise: Computability of Squared-error vs Absolute Error Estimators, (with discussion). Statistical Science, (1997) 12, 279-300.

Koenker, R. and Ng, P (2003) Inequality Constrained Quantile Regression, preprint.

See Also

qrisk

[Package quantreg version 5.94 Index]