rqProcess {quantreg}	R Documentation

Compute Standardized Quantile Regression Process

Description

Computes a standardize quantile regression process for the model specified by the formula, on the partition of [0,1] specified by the taus argument, and standardized according to the argument nullH. Intended for use in KhmaladzeTest.

Usage

rqProcess(formula, data, taus, nullH = "location", ...)

Arguments

formula
model formula

data
data frame to be used to interpret formula

taus
quantiles at which the process is to be evaluated, if any of the taus lie outside (0,1) then the full process is computed for all distinct solutions.

nullH
Null hypothesis to be used for standardization

...
optional arguments passed to summary.rq

Details

The process computes standardized estimates based on the hypothesis specified in the nullH argument. The Vhat component is rescaled by the Cholesky decomposition of the tau specific covariance matrix, the vhat component is rescaled by the marginal standard errors. The nature of the covariance matrix used for the standardization is controlled arguments passed via the ... argument to summary.rq. If the full process is estimated then these covariance options aren't available and only a simple iid-error form of the covariance matrix is used.

Value

taus
The points of evaluation of the process

qtaus
Values of xbar'betahat(taus)

Vhat
Joint parametric QR process

vhat
Marginal parametric QR processes

Author(s)

R. Koenker

See Also

KhmaladzeTest

[Package quantreg version 5.97 Index]