rqs.fit {quantreg}	R Documentation

Function to fit multiple response quantile regression models

Description

Function intended for multiple response quantile regression called from boot.rq for wild bootstrap option.

Usage

rqs.fit(x, y, tau=0.5, tol = 0.0001)

Arguments

x
the design matrix an n by p matrix.

y
the response variable as a n by m matrix

tau
the quantile desired, if tau lies outside (0,1)

tol
tolerance parameter for Barrodale and Roberts exterior point method.

See Also

boot.rq

[Package quantreg version 5.93 Index]